requirements_rl.txt
# Reinforcement Learning for Logistics Optimization Requirements
# Core ML and RL libraries
torch>=1.12.0
gym>=0.21.0
numpy>=1.21.0
pandas>=1.3.0

# Data processing and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Optimization and mathematical libraries
scipy>=1.7.0
scikit-learn>=1.0.0

# Geospatial calculations
shapely>=1.8.0

# Additional utilities
tqdm>=4.62.0
joblib>=1.1.0 